Convert tax_credits_eea_entitled_outcome outcome
I would've liked to have re-used the `tax_credits_going_abroad_helpline`
partial template for the Tax Credit Helpline contact details, but there are
a couple of small differences. I haven't been able to establish whether or
not the differences are intentional, so for now I've kept the contact details
in the outcome body template.