Disable session store
As requested by @tekin as a precautionary measure for all "front-end" apps.

The Smart Answers app is designed to be stateless - this change makes
that explicit.

Apparently the Varnish cache strips session cookies on the way in and out
anyway, so cookies aren't currently making it through to users, but at some
point such filters may be relaxed in order to implement authentication.
This change ensures that cookies will never make it through to users.

Also our cookie policy doesn't describe any of the cookies we might set, so
we'd be breaking our own policy by enabling sessions.

c.f. https://github.com/alphagov/government-frontend/pull/48

Although the configuration line was commented out in [1], this isn't sufficient
to disable session cookies; you need to actively set the session store to
`:disabled`.

Note that I was able to see the session cookie in the browser when the line was
just commented out, but only when I set a value in the session.

[1]: 84ba7be305c32abcee0fa89f48d5dd8c6a71bce0