Fix formatting of contacts in uk-benefits-abroad
Although the Govspeak docs [1] suggest that a contact block with generate HTML
line-break elements between each line, this doesn't appear to be the case (I
have raised a GitHub issue [2] asking about this).

It appears as if historically we've been working around this issue by adding
two trailing spaces to the lines within the contact blocks. This relies on a
feature of Kramdown which converts the two trailing spaces into HTML line break
elements [3].

Unfortunately the trailing spaces for the contact blocks in uk-benefits-abroad
were, almost certainly accidentally, lost in this commit [4].

Many developers have their editors set to automatically remove trailing
whitespace to make git history as clear as possible. I think this is generally
a good thing, but I suspect it may have been what led to these significant
trailing spaces being lost.

As a temporary solution, I've added two trailing backslashes to the relevant
lines. This is a Kramdown alternative to the two trailing spaces with the
advantage that hopefully they shouldn't get removed so easily by accident.

I think the correct longer-term solution is probably to fix Govspeak so that
it's behaviour matches the documentation, but since many projects use Govspeak,
it may take a while to get the change accepted.

[1]: https://github.com/alphagov/govspeak#contact
[2]: https://github.com/alphagov/govspeak/issues/55
[3]: http://kramdown.gettalong.org/syntax.html#paragraphs
[4]: 1a96852a3186254d03767f22b15754e5830200ac