DRY up next_steps of 'done' outcomes
The only difference between the highest_earner_done and husband_done outcomes
is the title.

I've introduced the done_body and done_next_steps partial in an attempt to DRY
these up, although I do wonder whether a better solution might be to have a
single outcome with a conditional in the `title`.