Make uk-benefits-abroad contact boxes consistent (surplus comma)
I missed this discrepancy between the contact boxes in the email conversation
with @lutgendorff and in the previous commit [1] where I tried to bring these
into line. I've agreed this change verbally with @lutgendorff just now -
apparently this is an "Oxford comma" type of scenario where "Monday to Friday"
should be followed by a comma, but just "Saturday" should not.

[1]: 56a93588e31e0378d02a57ec8c1802512f7f8db6