Avoid double-diffs in regression test assertions
Previously the `assert_equal` was diff-ing the output from `git diff` with an
empty string. This resulted in two levels of diff which made the output
confusing.

Asserting that the diff is empty and setting the diff as the assertion
failure message generates more sensible output for failing tests.

Fixes #1746.