Speed up common case of running regression tests for single flow
The `:preload_flows` option was being set to `true` which meant that **all**
smart answer flows were being loaded the first time the registries were
accessed i.e. in the first GET request. Unfortunately, even though we don't
currently run regression tests against them, the Smartdown smart answers take
approx 15 seconds to load on my machine at this point.

This commit sets the option to `false` which means that flows are loaded
on-demand, thus avoiding the 15 second penalty which is particularly
significant in the common use case of running the regression tests for a
single flow.

Unfortunately, because the registries do not cache the flows when not
pre-loading them, we incur a small performance hit, because the flow is
reloaded for every individual regression test combination. However, this
performance hit only becomes significant for the most complex flows like
`marriage-abroad`:

        towing-rules:       00m20s -> 00m05s (-15s)
        register-a-death:   00m35s -> 00m21s (-14s)
        uk-benefits-abroad: 00m51s -> 00m37s (-14s)
        marriage-abroad:    13m55s -> 14m30s (+35s)

I hesitated to make this change because of the latter performance hit, but
having chatted to @chrisroos, I decided that on balance this change is an
improvement.

If we're worried about the latter performance hit, I think the best solution
would be to change the behaviour of the registries to have more typical
caching behaviour i.e. even if the flows are not pre-loaded, the registry
should be able to cache the flow instance.