Unify countries that require 3 day residency and don't need birth certs
According to FCO countries that do not require a birth certificate as
a supporting document when issuing a CNI are the same countries that
require users to stay 3 days in the country before posting notice.

Italy is the only exception.

This commit is meant to make this connection more explicit.

Spain and Italy are added to the three_day_residency_requirement_applies
list, because they have this requirement, although it's handled by custom
phrases.