Convert amount_result outcome
I've had to introduce the `pays_reduced_ni_rate` precalculate block as it's
possible to get to this outcome without `pays_reduced_ni_rate` having been
defined. This didn't cause us any problems previously because the `State`
instances inherit from `OpenStruct` and will respond to any undefined method
with `nil`.

The amount_result.govspeak.erb template is pretty lengthy and can almost
certainly be DRYed up. I did try extracting the four main branches to partials
but ran into problems as I needed to pass locals in to each of the partials and
it wasn't obvious to me that the result was going to be much better than what I
have here.