Convert amount_result outcome
I think the only thing of any interest in here is the use of `if
defined?(pays_reduced_ni_rate)`. I have to use `defined?` as it turns out that
we can get to this branch of code in the template without
`pays_reduced_ni_rate` having been defined. This didn't cause us any problems
previously because the `State` instances inherit from `OpenStruct` and will
respond to any undefined method with `nil`.