Split single outcome into three outcomes
The `final_outcome` body is entirely made up of the `result_phrase` PhraseList.
Three versions of this PhraseList are built in three question nodes. It feels
simpler and more "idiomatic" to split the `final_outcome` into three outcomes,
representing the three paths by which a user could arrive there.

Also I'm hoping that this will make it easier to convert this flow to use
ERB templates for its outcomes.