Remove unused current_rules_national_insurance_and_state_pension
It's a bit hard to follow, but this phrase appears to have become redundant
in this commit [1], in which the `derived-entitlements` flow was renamed to
`state-pension-through-partner` at the same time as changes were made to the
flow. You can see that it was present before that commit [2], but missing
afterwards.

[1]: 0b69c3a8343d215d12dafe8ded058b344809381a
[2]: https://github.com/alphagov/smart-answers/blob/c92859936a1e2372fea394a56784e0cb8759814b/lib/flows/derived-entitlements.rb#L74
[3]: https://github.com/alphagov/smart-answers/blob/0b69c3a8343d215d12dafe8ded058b344809381a/lib/flows/state-pension-through-partner.rb#L73-L81