Split male_gender_outcome into two outcomes
These two partial templates always felt more like outcomes in their own right:

* impossibility_due_to_divorce
* impossibility_to_increase_pension

This commit increases the complexity of the logic in the flow and the number
of outcomes, but simplifies the logic in the templates themselves. I think
the result is easier to understand and more typical of other flows.