Convert you_can_use_result outcome
Some variables required by the new outcome template are only defined when the user has arrived at the outcome via a particular question.

When these variables were being interpolated via the I18n mechanism, they were accessed via accessor methods on the instance of `SmartAnswer::State`. The latter inherits from `OpenStruct` and so variables which were undefined would return `nil` by default.

However, the state variables are supplied to the ERB template as local variables. This means that a `ActionView::Template::Error: undefined local variable or method` exception is raised if an undefined variable is accessed in from the ERB template.

I've addressed this problem by adding explicit `precalculate` blocks in the new outcome which ensure that the relevant state variables exist. It's a bit ugly, but at least it makes the template's dependency on the outcome variables explicit.