Update calculate-your-redundancy-pay-questions-and-responses.yml
I started with arbitrary values and updated them to cover the Smart Answer flow.

This gives me 100% coverage of the redundancy_calculator.rb and all but 1 line
is covered in calculate-your-redundancy-pay. The only line not covered is the
`if years_employed < 2` branch in the `weekly_pay_before_tax?` question. This
is unreachable because the only way we can get to the `weekly_pay_before_tax?`
is if `years_employed` is greater than 2.