Split calculate-redundancy-pay.yml
This was the only file that contained text for multiple Smart Answers:

* calculate-employee-redundancy-pay
* calculate-your-redundancy-pay

I'm trying to add regression tests for calculate-your-redundancy-pay and the
generate-checksums-for-smart-anwer.rb script exploded because it expects a
one-to-one mapping between the Smart Answer flow and its YAML file.

Splitting the content of this file into two files, that match the Smart Answer
flows, allows me to add the regression tests and checksum data.