Convert contact_the_embassy outcome
This outcome was making use of the `overseas_passports_embassies` data partial.
I've implemented this by calling the standard `render partial: 'foo'`, but
specifying a relative path (`../data_partials/`) to the shared data partial
directory.

Ideally the data partial template would use the same file extension
(`.govspeak.erb`) as the outcome templates, but since a bunch of other flows
currently access via the old mechanism, I've decided to leave it be for now.