Convert oru_result outcome
This was the first time I've had to deal with a `PhraseList` element whose key
is constructed dynamically and where the key could take on a significant number
of different values i.e. `oru_courier_text_#{current_location}`.

I've chosen to address this using a `case` statement, but I'm not convinced
this is the best long-term approach.

The logic around this content is also somewhat split between the values in the
`when` clauses and the data in the `ORU_COURIER_VARIANTS` array in the data
query class. However, I don't think this split in the logic has got any worse
in that previously it was split between the YAML file and the data query class.

I've had to remove quite a few `assert_phrase_list` statements from the
integration tests, because the `PhraseList`'s no longer exist. In some cases
the test names are probably no longer appropriate, but I haven't addressed
that problem here.