Convert flat_rate_result outcome
I've duplicated the 'collect_and_pay_service' and 'fees_title' precalculate
blocks in the template. I considered changing the precaculate blocks so that
they returned strings (the values that were previously in the YAML files and
are now in the conditional in the ERB template) but felt uncomfortable doing
that: The strings feel like they belong in the template rather than the flow.

I've expanded the 'enforcement_charge' precalculate block in the ERB template.