Convert reduced_and_basic_rates_result outcome
As in the previous commit, I've duplicated the 'collect_and_pay_service' and
'fees_title' precalculate blocks in the ERB template. I got rid of these
precalculate blocks now that we're no longer using them.

I expanded the 'enforcement_charge' precalculate block in the ERB template and
removed it now that we're no longer using it.