Update CNI third country links
When a user wants to get married in a CNI country but is residing in a
third country, we show them two options: coming back to the UK or going
to the country where they intend to get married. For the former it used
to point to the uk_england outcome as a temporary measure. Now with
uk_residency question removed (previous commit) the links need to be updated
to point to the desired outcome.