Update copy refering to CNI acquired in the UK
Channel Islands and Isle of Man residents are now directed to
regular CNI outcomes for UK residents. Add links to authorities in Channel Islands
and Isle of Man, add a message about validity of CNIs issued in those parts
of the UK (it varies in all parts, so advise user to contact authorities).