Convert compressed_hours_done outcome
In moving the 'answer_compressed_hours' phrase from YAML to ERB I've converted
the apostrophe from "'" to "’". This is so that the rendered outcomes are
identical.

I don't understand why, but before making this change, the "'" was being
rendered as "&#8217;" whereas it's rendered as "’" when it comes from a string
in YAML.