Remove unused phrases from register-a-death
These phrase keys don't appear to be used anywhere and the regression tests
all pass without them, so I think it's safe to remove them. However, I thought
I'd put this in a pull request to get a second opinion.

I've updated the checksums to indicate that I'm happy that the removal of these
phrases have had no effect on the behaviour of the system.