Convert casual_or_irregular_hours_done outcome
In moving the 'answer_hours_minutes' phrase from YAML to ERB I've converted
the apostrophe from "'" to "’". This is so that the rendered outcomes are
identical.

I don't understand why, but before making this change, the "'" was being
rendered as "&#8217;" whereas it's rendered as "’" when it comes from a string
in YAML.

This was the last use of the 'your_employer_with_rounding' phrase so I've now
removed it.