Remove a test that became redundand after removing uk residency question
This test became duplicated (exactly the same inputs and assertions) as another
one defined in line 512.

The test was found by accumulating all responses in the marriage abroad
integration test suite  and checking if the same response set was added already.
It was added in flow_test_helper.rb in the code asserting current node. The script
is very hackish so I'm not going to add it here.