Convert outcome_study_m outcome
I've updated one of the rendered outcomes as it has one less newline than
before. I originally added the newline to the '_stateless_or_refugee' partial
but that results in different HTML rendered by the following templates:

* outcome_medical_y
* outcome_school_y
* outcome_marriage
* outcome_standard_visit

It doesn't seem to be possible to exactly replicate the existing behaviour of
the YAML template:

Given the following YAML snippet:

    body: |
      %{if_refugee}

      Apply for a [Short-term Study visa](/study-visit-visa) if you are
      studying for 6 months or less.

When 'if_refugee' is nil I get a single newline between the '.summary' div that
contains the title, and the "Apply for..." paragraph.

    <div class="summary"><h2 class="result-title">You’ll need a visa to study
    in the UK</h2></div>
    <p>Apply for a <a href="/study-visit-visa">Short-term Study visa</a> if
    you are studying for 6 months or less.</p>

When 'if_refugee' is a PhraseList I get two newlines between the '.summary' div
that contains the title, and the div containing the "You must apply for..."
paragraph.

    <div class="summary"><h2 class="result-title">You’ll need a visa to study
    in the UK</h2></div>

    <div role="note" aria-label="Information" class="application-notice
    info-notice">
    <p>You must apply for your visa from the country you&rsquo;re originally
    from or currently living in.</p>
    </div>

    <p>Apply for a <a href="/study-visit-visa">Short-term Study visa</a> if you
    are studying for 6 months or less.</p>

I don't seem to be able to replicate this in ERB. If I have a newline between
rendering the 'stateless_or_refugee' partial and displaying the "Apply for..."
paragraph then I get an additional newline in the 5 rendered outcomes that
don't render anything in the 'stateless_or_refugee' partial. If I don't have a
newline then I lose the newline in the 1 rendered outcome that renders the
'stateless_or_refugee' partial.