Use date String vs date Object in integration tests
I'm not sure this is hugely important, but it captures the changes originally
made [1] in a branch [2] I'm trying to rebase against `master`. The change is
unrelated to the other changes in that branch and making the changes directly
in `master` should make it easier to rebase the branch and make the
corresponding pull request more focussed.

I chose not to change the date question validation tests, because I wanted to
keep them consistent with the same type of tests elsewhere in the codebase.

[1]: 3f805b803b992756e0999e414fc42717177b00bd
[2]: statutory-sick-pay-updates