Fix typo in plan_adoption/maternity_leave unit test names
It looks as if the word "last" was copied and pasted form the previous
test name in each case.

In general, the names of these tests leave a lot to be desired, but I don't
want to get sucked into fixing them all right now, because that's a bigger job.