Format all dates for plan-adoption-leave in "calculator" methods
Prior to this commit some but not all dates & date ranges were being formatted
in methods on `SmartAnswer::Calculators::PlanAdoptionLeave`; while others were
being formatted using the `:long` format prescribed in
`NodePresenter#value_for_interpolation` [1] and defined in `en.yml` [2].

I'm about to convert the outcome to render using an ERB template and this means
that it will no longer be possible to make use of the default `NodePresenter`
behaviour to format dates as described above. I could've formatted the two
relevant dates within `precalculate` blocks in the flow, but since some were
already being formatted in `SmartAnswer::Calculators::PlanAdoptionLeave`, I
decided that the best option was to add methods to do the same for these two
relevant dates.

Although in the longer term I'd prefer to setup some `Date::DATE_FORMATS` and
even a default one to be used implicitly via `Date#to_s` within the template,
this seems like the best solution for now.

[1]: https://github.com/alphagov/smart-answers/blob/de7619dfd88c8f344c67285f003c0e6f159ec3ec/app/presenters/node_presenter.rb#L27
[2]: https://github.com/alphagov/smart-answers/blob/f9974a7a850d874f3bf38d1720efcaac8f1a61ef/config/locales/en.yml#L7