Convert ips_application_result_online outcome
* `birth_location` is saved as an input in the `country_of_birth?` question,
but it is possible to reach this outcome without visiting this question. So
I've added a `precalculate` block which ensures it's available to the ERB
template whether it's been set or not.

* I've introduced the `how_long` & `child_or_adult_passport_costs_ips` partial
templates, because I can see they're going to be needed for the
`ips_application_result` outcome.

* I've done a very literal conversion of dynamically constructed `PhraseList`
keys by making use of `case` statements. It's not particularly pretty but it'll
do for now.

* I've fairly blindly removed all the uses of `assert_phrase_list` in tests
associated with this outcome, because the phrase lists no longer exists; we're
now relying on the regression tests to give us coverage.

* I've left the assertions matching against `outcome_body` in place, because
these work with the ERB templates.