Do not allow linked sickness duration to be negative
Make sure users can not enter linked sickness end date that is before
the start day by mistake. Add a clear message in the UI:
'End date should be on or after start date'

The prior_sick_days calculation needed to become next_node_calculation
for the validation to be able to access its value.