Convert ips_application_result outcome
* `birth_location` is saved as an input in the `country_of_birth?` question,
but it is possible to reach this outcome without visiting this question. So
I've added a `precalculate` block which ensures it's available to the ERB
template whether it's been set or not.

* I've also added a `precalculate` block for the `data_query` local variable to
make it available to the template.

* I've done a very literal conversion of dynamically constructed `PhraseList`
keys by making use of `case` statements. It's not particularly pretty but it'll
do for now.

* While I've introduced quite a few partial templates, I've only/mostly done
this when a phrase list element is used in more than one place. In hindsight it
_might_ have been clearer to use partials as a structural device for more of
the phrase list elements to more closely match the original code in the flow
class.

* I've fairly blindly removed all the uses of `assert_phrase_list` in tests
associated with this outcome, because the phrase lists no longer exists; we're
now relying on the regression tests to give us coverage. In tests which ended
up with no assertions, I've chosen to introduce an `assert_current_node`
statement to give the test some meaning. I could probably equally well have
removed those tests.

* I've left the assertions matching against `outcome_body` in place, because
these work with the ERB templates.