Regenerate HTML outputs for minimum-wage-calculator-employers
These became out of sync in #1773 when changes were made to the underlying
`SmartAnswer::Calculators::MinimumWageCalculator` [1] - the HTML outputs for
`am-i-getting-minimum-wage` were updated at the time [2], but the ones for
`minimum-wage-calculator-employers` were missed.

Given that the HTML outputs for `am-i-getting-minimum-wage` haven't changed
since [2] and all the differences to the HTML outputs for
`minimum-wage-calculator-employers` seem to be the same figures changing,
I'm happy to accept the changes by committing them to `master`.

[1]: f30d478614bce1db6f4d739907ffa0da52e1682e
[2]: ad6efefefa6027bf1a636b9bbc796c5c0c8aa9d1