Share acas_information partial between the two minimum wage flows
The partial is identical, so it seems to make sense to share it.

I contemplated re-using the `data_partials` directory or the `shared_logic`
directory, but in the end I went with a new `shared` directory.

I used the `shared/minimum_wage` sub-directory, because the partial is
specific to the two minimum wage flows and not any other flows.