Remove unnecessary assert_state_variable assertions
It looks like these `assert_state_variable <var>, nil` assertions started to
appear in b9c1d714e748ec009ada55bfcc75e77a64c6f090.

My guess is that the intention was to ensure that these keys weren't appearing
in Phrase lists for the given test set-up. That wouldn't have worked because
adding a key to a Phrase list doesn't result in it being made available as a
state variable.

The following keys don't appear anywhere in the code:

* header_hot_water_cyclinder_jacket
* opt_cylinder_thermostat
* opt_external_doors
* opt_fan_assisted_heater
* opt_micro_chp
* opt_micro_wind
* opt_secondary_glazing
* opt_solar
* opt_warm_air_unit

The following keys appear in Phrase lists but it still doesn't make sense to
include them in calls to `assert_state_variable`:

* opt_biomass_boilers_heaters
* opt_cavity_wall_insulation
* opt_draught_proofing
* opt_heat_pump
* opt_loft_roof_insulation
* opt_replacement_glazing
* opt_solar_water_heating
* opt_solid_wall_insulation