Introduce text format for main `show` action
This is so we can generate the raw Govspeak output for outcome nodes. The idea
is to use this format to generate more manageable outputs from the regression
tests.

The action renders 404 Not Found when format is `text` for non-outcome nodes.

The new `text` format is disabled in production simply because we don't know
whether it will work correctly..