Rename variables that end with a question mark in paternity-calculator
Rename variables that end with a question mark

State variables are supplied in the `locals` hash to `ActionView::Base#render`
(in the `OutcomePresenter`) where they are converted to Ruby variables for use
in the template. While a hash can have a key that ends with a question mark,
the same is not true for Ruby variables. State variables ending in question
marks cause SyntaxErrors in ActionView:

    irb> template_path = Rails.root.join('app/views')
    irb> view = ActionView::Base.new([template_path])
    irb> template = 'smart_answers/show'
    irb> locals = {question?: 'value'}
    irb> view.render(template: template, locals: locals)

    SyntaxError: /app/views/smart_answers/show.html.erb:1:
    syntax error, unexpected '='
    ...er = @output_buffer;question? = question? = local_assigns[:q...