Remove redundant sentence in Kazakhstan
It used to read:

If you need a CNI, you’ll first need to give notice of your intended marriage (banns)
at the British Embassy Office in Almaty.

Make an appointment at the embassy in Almaty to give notice of your marriage.
There’s a fee for this service (see below).

Make an appointment at the embassy in Almaty.