Convert paternity_leave_and_pay outcome
This was a bit painful to convert, and highlighted two problems with some
existing outcomes in master:

Problem 1
---------

The `paternity_entitled_to_leave` phrase couldn't be rendered (see further down
this commit note for more information) which means that we end up with that
string ('paternity_entitled_to_leave') in certain outcomes. For example:

/maternity-paternity-calculator/y/adoption/no/2015-04-05/2015-04-06/yes/no/yes/2
015-04-06/2015-01-01/2014-11-01/every_4_weeks/3000.0/usual_paydates/2015-01-01

Problem 2
---------

The `adoption_spp_claim_link` phrase doesn't exist. This was being requested in
the `paternity_info` calculate block when `leave_type` was 'adoption'
(`"#{leave_type}_spp_claim_link"`). In master this results in the
'adoption_spp_claim_link' string being rendered but in this branch I'm not
rendering anything. I plan to take this up with the content team separately.

You can see an example of this problem at:

/maternity-paternity-calculator/y/adoption/no/2015-04-05/2015-04-06/yes/no/yes/2
015-04-06/2015-01-01/2014-11-01/every_4_weeks/3000.0/usual_paydates/2015-01-01

Changes
-------

I've updated the rendered outcomes that previously suffered from problems 1 and
2 listed above.

I inlined this outcome-specific phrase:

* paternity_entitled_to_pay

I've added partials for phrases that appear to be used elsewhere:

* paternity_not_entitled_to_leave
* paternity_entitled_to_leave
* paternity_not_entitled_to_pay_intro
* paternity_not_entitled_to_pay_outro

I've added the 'spp_claim_link' partial to replicate the
`"#{leave_type}_spp_claim_link"` Phrase. I've inlined the text for
'paternity_adoption' and 'paternity' leave types (the
'paternity_adoption_spp_claim_link' and 'paternity_spp_claim_link' keys
respectively). The previous implementation would return the phrase key when it
couldn't be found in YAML. This resulted in adoption_spp_claim_link being
displayed in the page, because that key doesn't exist. See Problem 2 in this
commit note.

I've updated the `entitled_to_pay` precalculate block as `paternity_info` no
longer exists. Checking whether `paternity_info` included the
`:paternity_entitled_to_pay` key is the same as checking whether
`above_lower_earning_limit` is true so I've changed the implementation to use
that instead.

I've had to add the following precalculate blocks for things that are required
by the template but not necessarily set during certain paths through the flow:

* has_contract
* leave_spp_claim_link
* notice_of_leave_deadline

I've had to update `format_date` to return nil when the value supplied is nil.
This is to replicate the behaviour of `NodePresenter#value_for_interpolation`
as that will return the value if it's not of a certain type.

Further info about Problem 1
----------------------------

Rendering the `paternity_entitled_to_leave` phrase fails because rendering
anything from the YAML file results in a call to `state_for_interpolation`.
This results in every method on the the `State` object being rendered. Some of
the methods on the State object will return PhraseLists that fail to render
because of missing interpolation values. If any of these fail then the primary
Phrase list will also fail.