Convert paternity_not_entitled_to_leave_or_pay outcome
I've inlined these outcome-specific phrases:

* paternity_adoption_must_be_employed_by_you
* paternity_must_be_employed_by_you
* paternity_adoption_not_responsible_for_upbringing
* paternity_not_responsible_for_upbringing
* paternity_not_entitled_to_leave_or_pay_intro
* paternity_not_entitled_to_leave_or_pay_outro

I've removed these phrases that are already in partials and no longer used
elsewhere:

* paternity_entitled_to_leave
* paternity_not_entitled_to_leave
* paternity_not_entitled_to_pay_intro
* paternity_not_entitled_to_pay_outro

I've added partials for these phrases that appear to be used elsewhere:

* not_worked_long_enough

I've added precalculate blocks to the outcome that are required by the
template. It appears to be possible to get to this outcome without these
variables having been set.

* has_contract
* paternity_employment_start