Validate linked sickness start date is before the recent sickness
Linked sickness by definition is a sickness withing 8 weeks in the past
from the recent sickness start date. If it is in future it is not a linked
sickness, perhaps users need to swap the two date ranges.

There were two tests that did not satisfy this condition. I have changed the
inputs as I believe they are incorrect.