Always expose cost_change_4_weeks to views
This is addressing an error from errbit.
Sample affected path: /childcare-costs-for-tax-credits/y/yes/yes/monthly_diff_amount/2400.0/127.0

cost_change_4_weeks was only set if it's true, otherwise erb views could not
see it as a defined variable. There was also no input set for regression tests
that covers a case where `ten_or_more` is true, therefore this code branch
in erb was never exected.

The issue is fixed by explicitly exposing cost_change_4_weeks for the outcome
that may need it and setting it to false unless it's already set.

Add a regression test input to reproduce the issue and some extra assertions
to highlight the importance of cost_change_4_weeks.