Remove 'local' from 'Make an appointment at the local British Embassy'
The word is redundand. We provide an appointment booking link or the
embassy address below this sentence. Historically we did not always have
the country name in this sentence, therefore the word 'local' used to indicate
that it's the embassy in the ceremony country. Now we explicitly say
'embassy in %{country name}'.