Reuse document return fees table for births and deaths
Reuse nearly identical dcument return fees table for births and
deaths registration smartanswers by exracting it to a partial.

The only stylistic difference the use of `<tr>` instead of `<th>`
for table headers in register-a-death. Both use `<th>` now. The
table above the document return fees table in register-a-death is
also using `<th>` for the header now.